Update used for CPTAC3 catalog3 run list development

+++ src/make_canonical_run_list.py
@@ -11,21 +11,22 @@ def eprint(*args, **kwargs):
     print(*args, file=sys.stderr, **kwargs)
 
 def read_catalog(catalog_fn):
-    catalog = pd.read_csv(catalog_fn, sep="\t")
+    catalog = pd.read_csv(catalog_fn, sep="\t", keep_default_na=False)
     # at some point, expand metadata into something convenient
     catalog['metadata'] = catalog.apply(lambda row: json.loads(row['metadata']), axis=1)
     return catalog
 
 
 # dataset columns: ['dataset_name', 'uuid', 'case', 'disease', 'aliquot_tag']
-def get_dataset_list(catalog, cases, sample_types, alignment, experimental_strategy, data_format, data_variety):
+def get_dataset_list(catalog, cases, sample_types, alignment, experimental_strategy, data_format, data_variety, debug):
     case_loc = catalog['case'].isin(cases)
     sample_types_loc = catalog['sample_type'].isin(sample_types)
 
-#    eprint("alignment = " + str(alignment))
-#    eprint("experimental_strategy = " + str(experimental_strategy))
-#    eprint("data_format = " + str(data_format))
-#    eprint("data_variety = " + str(data_variety))
+    if debug:
+        eprint("alignment = " + str(alignment))
+        eprint("experimental_strategy = " + str(experimental_strategy))
+        eprint("data_format = " + str(data_format))
+        eprint("data_variety = " + str(data_variety))
 
 # https://stackoverflow.com/questions/17071871/how-do-i-select-rows-from-a-dataframe-based-on-column-values
     alignment_loc = (catalog['alignment'] == alignment) if alignment is not None else True
@@ -41,9 +42,11 @@ def get_dataset_list(catalog, cases, sample_types, alignment, experimental_strat
     return( catalog.loc[all_loc, ['dataset_name', 'uuid', 'case', 'disease', 'aliquot_tag']])
 
 # run_name is the case name with aliquot tags appended for any datasets with multiplicity > 1
-def get_run_name(case, aliquot1_tag, multiples_ds1, aliquot2_tag = None, multiples_ds2 = None):
+def get_run_name(case, run_suffix, aliquot1_tag, multiples_ds1, aliquot2_tag = None, multiples_ds2 = None):
     # ds columns: ['dataset_name', 'uuid', 'case', 'disease', 'aliquot_tag']])
     run_name = case
+    if run_suffix is not None:
+        run_name = run_name + '.' + run_suffix
     if multiples_ds1 > 1:
         run_name = run_name + '.' + aliquot1_tag
     if aliquot2_tag is not None:
@@ -63,10 +66,11 @@ def get_run_name(case, aliquot1_tag, multiples_ds1, aliquot2_tag = None, multipl
 # this works only for one case right now
 # example rows of dlm
 # ['C3L-00017.WXS.T.hg38', '4e2c5edf-8162-46f2-bb3e-11de6846c0e3', 'C3L-00017', 'PDA', 'ALQ_be7244ce']
-def get_paired_run_list(dl1, dl2, pipeline_data):
+def get_paired_run_list(dl1, dl2, pipeline_info):
     multiples_ds1 = dl1.shape[0]
     multiples_ds2 = dl2.shape[0]
 
+    suffix = pipeline_info['run_suffix'] if 'run_suffix' in pipeline_info else None
 # https://stackoverflow.com/questions/45672342/create-a-dataframe-of-permutations-in-pandas-from-list
     dlm=list(itertools.product(dl1.values.tolist(), dl2.values.tolist()))
 
@@ -76,7 +80,7 @@ def get_paired_run_list(dl1, dl2, pipeline_data):
         ds1 = d[0]
         ds2 = d[1]
 
-        run_name = get_run_name(ds1[2], ds1[4], multiples_ds1, ds2[4], multiples_ds2)
+        run_name = get_run_name(ds1[2], suffix, ds1[4], multiples_ds1, ds2[4], multiples_ds2)
         # note that run_metadata currently has value of case.  This needs to be updated
         row={"run_name": run_name, "run_metadata": ds1[2], "dataset1_name": ds1[0], "dataset1_uuid": ds1[1], "dataset2_name": ds2[0], "dataset2_uuid":  ds2[1]}
 
@@ -98,19 +102,20 @@ def get_paired_run_list(dl1, dl2, pipeline_data):
 #   * label1 - common name for dataset1, e.g. "tumor" (and label2 would be "normal")
 
 # NOTE: currently, run_metadata is not functional
-def get_single_run_list(dl, pipeline_data):
+def get_single_run_list(dl, pipeline_info):
 #    run_list is catalog.loc[all_loc, ['dataset_name', 'uuid', 'case', 'disease', 'aliquot_tag']]
     dl = dl.rename(columns={'uuid': 'dataset_uuid'})
     multiples_ds1 = dl.shape[0]
+    suffix = pipeline_info['run_suffix']
 
     #dl['run_name'] = dl.apply(lambda row: get_run_name(row, multiples_ds1), axis=1 )
-    dl['run_name'] = dl.apply(lambda row: get_run_name(row['case'], row['aliquot_tag'], multiples_ds1), axis=1 )
+    dl['run_name'] = dl.apply(lambda row: get_run_name(row['case'], suffix, row['aliquot_tag'], multiples_ds1), axis=1 )
 
     # run_metadata is not implemented.
     # for now, run_metadata is simply the case name
 
-    # create json string based on run_metadata and pipeline_data information
-    #dl['run_metadata'] = dl.apply(lambda row: json.dumps(run_metadata.update(pipeline_data)), axis=1 )
+    # create json string based on run_metadata and pipeline_info information
+    #dl['run_metadata'] = dl.apply(lambda row: json.dumps(run_metadata.update(pipeline_info)), axis=1 )
     dl = dl.rename(columns={'case': 'run_metadata'})
 
     return dl[['run_name', 'run_metadata', 'dataset_name', 'dataset_uuid']].reset_index(drop=True)
@@ -120,7 +125,7 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Generate run list for cases of interest from Catalog3 file for single and paired runs ")
     parser.add_argument("-d", "--debug", action="store_true", help="Print debugging information to stderr")
     parser.add_argument("-C", "--catalog", dest="catalog_fn", help="Catalog file name", required=True)
-    parser.add_argument("-o", "--output", dest="outfn", default="stdout", help="Output file name.  Default writes to stdout")
+    parser.add_argument("-o", "--output", dest="outfn", default="stdout", help="Output file name.  Appends to file if it exists.  Default writes to stdout")
     parser.add_argument("-a", "--alignment", help="Alignment of datasets, e.g., 'harmonized'")
     parser.add_argument("-e", "--experimental_strategy", help="Experimental strategy of datasets, e.g., 'WGS'")
     parser.add_argument("-f", "--data_format", help="Data format of datasets, e.g., 'BAM'")
@@ -130,6 +135,7 @@ if __name__ == "__main__":
     parser.add_argument("-T", "--sample_type2", help="Comma-separated list of sample types for sample2.  Implies paired workflow")
     parser.add_argument("-l", "--label1", default="dataset1", help="Label used for this dataset, e.g., 'tumor'")
     parser.add_argument("-L", "--label2", default="dataset2", help="Label used for this dataset, e.g., 'normal'")
+    parser.add_argument("-R", "--run_suffix", help="Additional string to be added to run_name following case name")
     parser.add_argument("-p", "--pipeline", help="Target pipeline name")
     parser.add_argument('cases', nargs='+', help="Cases to be evaluated")
 
@@ -153,26 +159,35 @@ if __name__ == "__main__":
         pipeline_info.update({'target_pipeline': args.pipeline})
     if args.label1:
         pipeline_info.update({'label1': args.label1})
+    if args.run_suffix:
+        pipeline_info.update({'run_suffix': args.run_suffix})
 
     # it is easier for a pandas newbie to iterate over all cases rather than process the list whole
     # get_dataset_list works OK with a list of cases
     read_list = None
     for case in args.cases:
-        dataset_list1 = get_dataset_list(catalog, [case], sample_types, args.alignment, args.experimental_strategy, args.data_format, args.data_variety)
+        dataset_list1 = get_dataset_list(catalog, [case], sample_types, args.alignment, args.experimental_strategy, args.data_format, args.data_variety, args.debug)
 
         if is_paired_workflow:
-            dataset_list2 = get_dataset_list(catalog, [case], sample_types2, args.alignment, args.experimental_strategy, args.data_format, data_variety2)
+            dataset_list2 = get_dataset_list(catalog, [case], sample_types2, args.alignment, args.experimental_strategy, args.data_format, data_variety2, args.debug)
             rl = get_paired_run_list(dataset_list1, dataset_list2, pipeline_info)
         else:
             rl = get_single_run_list(dataset_list1, pipeline_info)
 
         read_list = read_list.append(rl) if read_list is not None else rl
 
+    write_header = True
     if args.outfn == "stdout":
         o = sys.stdout
     else:
-        print("Writing catalog to " + args.outfn)
-        o = open(args.outfn, "w")
+        # Check if output file exists.  If it does, append to it
+        if os.path.isfile(args.outfn):
+            print("Appending run_list to " + args.outfn)
+            o = open(args.outfn, "a")
+            write_header = False
+        else:             
+            print("Writing run_list to " + args.outfn)
+            o = open(args.outfn, "w")
 
-    read_list.to_csv(o, sep="\t", index=False)
+    read_list.to_csv(o, sep="\t", index=False, header=write_header)
